FTP tested and complete
Disclaimer: This is just a prototype and is really really sketchy,
Only supports the most basic functionalty and will not retain state if crashed. The program keeps track of all the files in a hashtable and will be lost if an error is occured and the server will not be able to serve those files anymore.

Python is next to a potato when it comes to multiptrocessing. Im using a library for the HTTP server and its really difficult to pass objects that are to be shared with processes. After spending a day on this I reverted to a single processsd (and single threaded) approach. This will still work for multiple simultaneous requests but will serve 1 at a time.

Once we are done testing the major components, we can rebuild this part in JAVA or C++ ( Calling dibs on this)

Read the howto.txt and the protocol file to understand how to upload and download files